完成了分页和查询，还有导出和验证码未完成，对于分页的代码些还没完全搞懂，明天研究一下逻辑

+++ src/com/course/utils/ExcelRead.java
@@ -12,9 +12,8 @@ import org.apache.poi.ss.usermodel.WorkbookFactory;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+
 
 /**
 *   文本工具类： 文本工具类：
@@ -42,11 +41,11 @@ public class ExcelRead {
                 for (int i = 0; i <= rowNum ; i++) {
                     ArrayList<String> courseString =new ArrayList<String>();
                     Row row = sheet.getRow(i);
-                    courseString.add(String.valueOf(row.getCell(0).getNumericCellValue()));
+                    courseString.add(String.valueOf((int)row.getCell(0).getNumericCellValue()));
                     courseString.add(row.getCell(1).getStringCellValue());
                     courseString.add(row.getCell(2).getStringCellValue());
                     courseString.add(row.getCell(3).getStringCellValue());
-                    courseString.add(String.valueOf(row.getCell(4).getNumericCellValue()));
+                    courseString.add(String.valueOf((int)row.getCell(4).getNumericCellValue()));
                     courseString.add(row.getCell(5).getStringCellValue());
                     result.add(courseString);
                 }
